Added styling for the new multi-selection toggle button
This update fine-tunes the styling for the new multi-select accordion. It makes sure that when you open several questions at once, they all look good and are easy to read. The way answers show up has been adjusted to work smoothly with this feature. Finally, the toggle button has been designed to match the background smoothly with a change when covered + animation.